Struct google_dfareporting3::api::Creative

pub struct Creative {
    pub account_id: Option<String>,
    pub active: Option<bool>,
    pub ad_parameters: Option<String>,
    pub ad_tag_keys: Option<Vec<String>>,
    pub advertiser_id: Option<String>,
    pub allow_script_access: Option<bool>,
    pub archived: Option<bool>,
    pub artwork_type: Option<String>,
    pub authoring_source: Option<String>,
    pub authoring_tool: Option<String>,
    pub auto_advance_images: Option<bool>,
    pub background_color: Option<String>,
    pub backup_image_click_through_url: Option<CreativeClickThroughUrl>,
    pub backup_image_features: Option<Vec<String>>,
    pub backup_image_reporting_label: Option<String>,
    pub backup_image_target_window: Option<TargetWindow>,
    pub click_tags: Option<Vec<ClickTag>>,
    pub commercial_id: Option<String>,
    pub companion_creatives: Option<Vec<String>>,
    pub compatibility: Option<Vec<String>>,
    pub convert_flash_to_html5: Option<bool>,
    pub counter_custom_events: Option<Vec<CreativeCustomEvent>>,
    pub creative_asset_selection: Option<CreativeAssetSelection>,
    pub creative_assets: Option<Vec<CreativeAsset>>,
    pub creative_field_assignments: Option<Vec<CreativeFieldAssignment>>,
    pub custom_key_values: Option<Vec<String>>,
    pub dynamic_asset_selection: Option<bool>,
    pub exit_custom_events: Option<Vec<CreativeCustomEvent>>,
    pub fs_command: Option<FsCommand>,
    pub html_code: Option<String>,
    pub html_code_locked: Option<bool>,
    pub id: Option<String>,
    pub id_dimension_value: Option<DimensionValue>,
    pub kind: Option<String>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub latest_trafficked_creative_id: Option<String>,
    pub name: Option<String>,
    pub override_css: Option<String>,
    pub polite_load_asset_id: Option<String>,
    pub progress_offset: Option<VideoOffset>,
    pub redirect_url: Option<String>,
    pub rendering_id: Option<String>,
    pub rendering_id_dimension_value: Option<DimensionValue>,
    pub required_flash_plugin_version: Option<String>,
    pub required_flash_version: Option<i32>,
    pub size: Option<Size>,
    pub skip_offset: Option<VideoOffset>,
    pub skippable: Option<bool>,
    pub ssl_compliant: Option<bool>,
    pub ssl_override: Option<bool>,
    pub studio_advertiser_id: Option<String>,
    pub studio_creative_id: Option<String>,
    pub studio_trafficked_creative_id: Option<String>,
    pub subaccount_id: Option<String>,
    pub third_party_backup_image_impressions_url: Option<String>,
    pub third_party_rich_media_impressions_url: Option<String>,
    pub third_party_urls: Option<Vec<ThirdPartyTrackingUrl>>,
    pub timer_custom_events: Option<Vec<CreativeCustomEvent>>,
    pub total_file_size: Option<String>,
    pub type_: Option<String>,
    pub universal_ad_id: Option<UniversalAdId>,
    pub version: Option<i32>,
    pub video_description: Option<String>,
    pub video_duration: Option<f32>,
}

Contains properties of a Creative.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get creatives (response)

insert creatives (request|response)

list creatives (none)

patch creatives (request|response)

update creatives (request|response)

Fields
account_id: Option<String>

Account ID of this creative. This field, if left unset, will be auto-generated for both insert and update operations. Applicable to all creative types.

active: Option<bool>

Whether the creative is active. Applicable to all creative types.

ad_parameters: Option<String>

Ad parameters user for VPAID creative. This is a read-only field. Applicable to the following creative types: all VPAID.

ad_tag_keys: Option<Vec<String>>

Keywords for a Rich Media creative. Keywords let you customize the creative settings of a Rich Media ad running on your site without having to contact the advertiser. You can use keywords to dynamically change the look or functionality of a creative. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

advertiser_id: Option<String>

Advertiser ID of this creative. This is a required field. Applicable to all creative types.

allow_script_access: Option<bool>

Whether script access is allowed for this creative. This is a read-only and deprecated field which will automatically be set to true on update. Applicable to the following creative types: FLASH_INPAGE.

archived: Option<bool>

Whether the creative is archived. Applicable to all creative types.

artwork_type: Option<String>

Type of artwork used for the creative. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

authoring_source: Option<String>

Source application where creative was authored. Presently, only DBM authored creatives will have this field set. Applicable to all creative types.

authoring_tool: Option<String>

Authoring tool for HTML5 banner creatives. This is a read-only field. Applicable to the following creative types: HTML5_BANNER.

auto_advance_images: Option<bool>

Whether images are automatically advanced for image gallery creatives. Applicable to the following creative types: DISPLAY_IMAGE_GALLERY.

background_color: Option<String>

The 6-character HTML color code, beginning with #, for the background of the window area where the Flash file is displayed. Default is white. Applicable to the following creative types: FLASH_INPAGE.

backup_image_click_through_url: Option<CreativeClickThroughUrl>

Click-through URL for backup image. Applicable to ENHANCED_BANNER when the primary asset type is not HTML_IMAGE.

backup_image_features: Option<Vec<String>>

List of feature dependencies that will cause a backup image to be served if the browser that serves the ad does not support them. Feature dependencies are features that a browser must be able to support in order to render your HTML5 creative asset correctly. This field is initially auto-generated to contain all features detected by Campaign Manager for all the assets of this creative and can then be modified by the client. To reset this field, copy over all the creativeAssets' detected features. Applicable to the following creative types: HTML5_BANNER. Applicable to DISPLAY when the primary asset type is not HTML_IMAGE.

backup_image_reporting_label: Option<String>

Reporting label used for HTML5 banner backup image. Applicable to the following creative types: DISPLAY when the primary asset type is not HTML_IMAGE.

backup_image_target_window: Option<TargetWindow>

Target window for backup image. Applicable to the following creative types: FLASH_INPAGE and HTML5_BANNER. Applicable to DISPLAY when the primary asset type is not HTML_IMAGE.

click_tags: Option<Vec<ClickTag>>

Click tags of the creative. For DISPLAY, FLASH_INPAGE, and HTML5_BANNER creatives, this is a subset of detected click tags for the assets associated with this creative. After creating a flash asset, detected click tags will be returned in the creativeAssetMetadata. When inserting the creative, populate the creative clickTags field using the creativeAssetMetadata.clickTags field. For DISPLAY_IMAGE_GALLERY creatives, there should be exactly one entry in this list for each image creative asset. A click tag is matched with a corresponding creative asset by matching the clickTag.name field with the creativeAsset.assetIdentifier.name field. Applicable to the following creative types: DISPLAY_IMAGE_GALLERY, FLASH_INPAGE, HTML5_BANNER. Applicable to DISPLAY when the primary asset type is not HTML_IMAGE.

commercial_id: Option<String>

Industry standard ID assigned to creative for reach and frequency. Applicable to INSTREAM_VIDEO_REDIRECT creatives.

companion_creatives: Option<Vec<String>>

List of companion creatives assigned to an in-Stream video creative. Acceptable values include IDs of existing flash and image creatives. Applicable to the following creative types: all VPAID and all INSTREAM_VIDEO with dynamicAssetSelection set to false.

compatibility: Option<Vec<String>>

Compatibilities associated with this creative. This is a read-only field. DISPLAY and DISPLAY_INTERSTITIAL refer to rendering either on desktop or on mobile devices or in mobile apps for regular or interstitial ads, respectively. APP and APP_INTERSTITIAL are for rendering in mobile apps. Only pre-existing creatives may have these compatibilities since new creatives will either be assigned DISPLAY or DISPLAY_INTERSTITIAL instead. IN_STREAM_VIDEO refers to rendering in in-stream video ads developed with the VAST standard. Applicable to all creative types.

Acceptable values are:

"APP"

"APP_INTERSTITIAL"

"IN_STREAM_VIDEO"

"DISPLAY"

"DISPLAY_INTERSTITIAL"

convert_flash_to_html5: Option<bool>

Whether Flash assets associated with the creative need to be automatically converted to HTML5. This flag is enabled by default and users can choose to disable it if they don't want the system to generate and use HTML5 asset for this creative. Applicable to the following creative type: FLASH_INPAGE. Applicable to DISPLAY when the primary asset type is not HTML_IMAGE.

counter_custom_events: Option<Vec<CreativeCustomEvent>>

List of counter events configured for the creative. For DISPLAY_IMAGE_GALLERY creatives, these are read-only and auto-generated from clickTags. Applicable to the following creative types: DISPLAY_IMAGE_GALLERY, all RICH_MEDIA, and all VPAID.

creative_asset_selection: Option<CreativeAssetSelection>

Required if dynamicAssetSelection is true.

creative_assets: Option<Vec<CreativeAsset>>

Assets associated with a creative. Applicable to all but the following creative types: INTERNAL_REDIRECT, INTERSTITIAL_INTERNAL_REDIRECT, and REDIRECT

creative_field_assignments: Option<Vec<CreativeFieldAssignment>>

Creative field assignments for this creative. Applicable to all creative types.

custom_key_values: Option<Vec<String>>

Custom key-values for a Rich Media creative. Key-values let you customize the creative settings of a Rich Media ad running on your site without having to contact the advertiser. You can use key-values to dynamically change the look or functionality of a creative. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

dynamic_asset_selection: Option<bool>

Set this to true to enable the use of rules to target individual assets in this creative. When set to true creativeAssetSelection must be set. This also controls asset-level companions. When this is true, companion creatives should be assigned to creative assets. Learn more. Applicable to INSTREAM_VIDEO creatives.

exit_custom_events: Option<Vec<CreativeCustomEvent>>

List of exit events configured for the creative. For DISPLAY and DISPLAY_IMAGE_GALLERY creatives, these are read-only and auto-generated from clickTags, For DISPLAY, an event is also created from the backupImageReportingLabel. Applicable to the following creative types: DISPLAY_IMAGE_GALLERY, all RICH_MEDIA, and all VPAID. Applicable to DISPLAY when the primary asset type is not HTML_IMAGE.

fs_command: Option<FsCommand>

OpenWindow FSCommand of this creative. This lets the SWF file communicate with either Flash Player or the program hosting Flash Player, such as a web browser. This is only triggered if allowScriptAccess field is true. Applicable to the following creative types: FLASH_INPAGE.

html_code: Option<String>

HTML code for the creative. This is a required field when applicable. This field is ignored if htmlCodeLocked is true. Applicable to the following creative types: all CUSTOM, FLASH_INPAGE, and HTML5_BANNER, and all RICH_MEDIA.

html_code_locked: Option<bool>

Whether HTML code is generated by Campaign Manager or manually entered. Set to true to ignore changes to htmlCode. Applicable to the following creative types: FLASH_INPAGE and HTML5_BANNER.

id: Option<String>

ID of this creative. This is a read-only, auto-generated field. Applicable to all creative types.

id_dimension_value: Option<DimensionValue>

Dimension value for the ID of this creative. This is a read-only field. Applicable to all creative types.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#creative".

last_modified_info: Option<LastModifiedInfo>

Creative last modification information. This is a read-only field. Applicable to all creative types.

latest_trafficked_creative_id: Option<String>

Latest Studio trafficked creative ID associated with rich media and VPAID creatives. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

name: Option<String>

Name of the creative. This is a required field and must be less than 256 characters long. Applicable to all creative types.

override_css: Option<String>

Override CSS value for rich media creatives. Applicable to the following creative types: all RICH_MEDIA.

polite_load_asset_id: Option<String>

The asset ID of the polite load image asset. Applicable to the creative type: DISPLAY.

progress_offset: Option<VideoOffset>

Amount of time to play the video before counting a view. Applicable to the following creative types: all INSTREAM_VIDEO.

redirect_url: Option<String>

URL of hosted image or hosted video or another ad tag. For INSTREAM_VIDEO_REDIRECT creatives this is the in-stream video redirect URL. The standard for a VAST (Video Ad Serving Template) ad response allows for a redirect link to another VAST 2.0 or 3.0 call. This is a required field when applicable. Applicable to the following creative types: DISPLAY_REDIRECT, INTERNAL_REDIRECT, INTERSTITIAL_INTERNAL_REDIRECT, and INSTREAM_VIDEO_REDIRECT

rendering_id: Option<String>

ID of current rendering version. This is a read-only field. Applicable to all creative types.

rendering_id_dimension_value: Option<DimensionValue>

Dimension value for the rendering ID of this creative. This is a read-only field. Applicable to all creative types.

required_flash_plugin_version: Option<String>

The minimum required Flash plugin version for this creative. For example, 11.2.202.235. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

required_flash_version: Option<i32>

The internal Flash version for this creative as calculated by Studio. This is a read-only field. Applicable to the following creative types: FLASH_INPAGE all RICH_MEDIA, and all VPAID. Applicable to DISPLAY when the primary asset type is not HTML_IMAGE.

size: Option<Size>

Size associated with this creative. When inserting or updating a creative either the size ID field or size width and height fields can be used. This is a required field when applicable; however for IMAGE, FLASH_INPAGE creatives, and for DISPLAY creatives with a primary asset of type HTML_IMAGE, if left blank, this field will be automatically set using the actual size of the associated image assets. Applicable to the following creative types: DISPLAY, DISPLAY_IMAGE_GALLERY, FLASH_INPAGE, HTML5_BANNER, IMAGE, and all RICH_MEDIA.

skip_offset: Option<VideoOffset>

Amount of time to play the video before the skip button appears. Applicable to the following creative types: all INSTREAM_VIDEO.

skippable: Option<bool>

Whether the user can choose to skip the creative. Applicable to the following creative types: all INSTREAM_VIDEO and all VPAID.

ssl_compliant: Option<bool>

Whether the creative is SSL-compliant. This is a read-only field. Applicable to all creative types.

ssl_override: Option<bool>

Whether creative should be treated as SSL compliant even if the system scan shows it's not. Applicable to all creative types.

studio_advertiser_id: Option<String>

Studio advertiser ID associated with rich media and VPAID creatives. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

studio_creative_id: Option<String>

Studio creative ID associated with rich media and VPAID creatives. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

studio_trafficked_creative_id: Option<String>

Studio trafficked creative ID associated with rich media and VPAID creatives. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

subaccount_id: Option<String>

Subaccount ID of this creative. This field, if left unset, will be auto-generated for both insert and update operations. Applicable to all creative types.

third_party_backup_image_impressions_url: Option<String>

Third-party URL used to record backup image impressions. Applicable to the following creative types: all RICH_MEDIA.

third_party_rich_media_impressions_url: Option<String>

Third-party URL used to record rich media impressions. Applicable to the following creative types: all RICH_MEDIA.

third_party_urls: Option<Vec<ThirdPartyTrackingUrl>>

Third-party URLs for tracking in-stream video creative events. Applicable to the following creative types: all INSTREAM_VIDEO and all VPAID.

timer_custom_events: Option<Vec<CreativeCustomEvent>>

List of timer events configured for the creative. For DISPLAY_IMAGE_GALLERY creatives, these are read-only and auto-generated from clickTags. Applicable to the following creative types: DISPLAY_IMAGE_GALLERY, all RICH_MEDIA, and all VPAID. Applicable to DISPLAY when the primary asset is not HTML_IMAGE.

total_file_size: Option<String>

Combined size of all creative assets. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

type_: Option<String>

Type of this creative. This is a required field. Applicable to all creative types.

Note: FLASH_INPAGE, HTML5_BANNER, and IMAGE are only used for existing creatives. New creatives should use DISPLAY as a replacement for these types.

universal_ad_id: Option<UniversalAdId>

A Universal Ad ID as per the VAST 4.0 spec. Applicable to the following creative types: INSTREAM_VIDEO and VPAID.

version: Option<i32>

The version number helps you keep track of multiple versions of your creative in your reports. The version number will always be auto-generated during insert operations to start at 1. For tracking creatives the version cannot be incremented and will always remain at 1. For all other creative types the version can be incremented only by 1 during update operations. In addition, the version will be automatically incremented by 1 when undergoing Rich Media creative merging. Applicable to all creative types.

video_description: Option<String>

Description of the video ad. Applicable to the following creative types: all INSTREAM_VIDEO and all VPAID.

video_duration: Option<f32>

Creative video duration in seconds. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO, all RICH_MEDIA, and all VPAID.