Struct google_dfareporting3::api::PlatformTypeGetCall

pub struct PlatformTypeGetCall<'a, C> where
    C: 'a,  { /* fields omitted */ }

Gets one platform type by ID.

A builder for the get method supported by a platformType resource. It is not used directly, but through a PlatformTypeMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.platform_types().get("profileId", "id")
             .doit().await;

Implementations

impl<'a, C> PlatformTypeGetCall<'a, C> where
C: BorrowMut<Client<HttpsConnector<HttpConnector>, Body>>,

pub async fn doit(self) -> Result<(Response<Body>, PlatformType)>

Perform the operation you have build so far.

pub fn profile_id(self, new_value: &str) -> PlatformTypeGetCall<'a, C>

User profile ID associated with this request.

Sets the profile id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn id(self, new_value: &str) -> PlatformTypeGetCall<'a, C>

Platform type ID.

Sets the id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn delegate(
self,
new_value: &'a mut dyn Delegate
) -> PlatformTypeGetCall<'a, C>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(self, name: T, value: T) -> PlatformTypeGetCall<'a, C> where
T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

alt (query-string) - Data format for the response.

fields (query-string) - Selector specifying which fields to include in a partial response.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

oauth_token (query-string) - OAuth 2.0 token for the current user.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

quotaUser (query-string) - An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

userIp (query-string) - Deprecated. Please use quotaUser instead.

pub fn add_scope<T, S>(self, scope: T) -> PlatformTypeGetCall<'a, C> where
T: Into<Option<S>>,
S: AsRef<str>,

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Dfatrafficking.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope. If None is specified, then all scopes will be removed and no default scope will be used either. In that case, you have to specify your API-key using the key parameter (see the param() function for details).

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.