Enum google_dfareporting3::Error

pub enum Error {
    HttpError(Error),
    UploadSizeLimitExceeded(u64, u64),
    BadRequest(ErrorResponse),
    MissingAPIKey,
    MissingToken(Box<dyn Error>),
    Cancelled,
    FieldClash(&'static str),
    JsonDecodeError(String, Error),
    Failure(Response),
}

Variants
HttpError(Error)

The http connection failed

UploadSizeLimitExceeded(u64, u64)

An attempt was made to upload a resource with size stored in field .0 even though the maximum upload size is what is stored in field .1.

BadRequest(ErrorResponse)

Represents information about a request that was not understood by the server. Details are included.

MissingAPIKey

We needed an API key for authentication, but didn't obtain one. Neither through the authenticator, nor through the Delegate.

MissingToken(Box<dyn Error>)

We required a Token, but didn't get one from the Authenticator

Cancelled

The delgate instructed to cancel the operation

FieldClash(&'static str)

An additional, free form field clashed with one of the built-in optional ones

JsonDecodeError(String, Error)

Shows that we failed to decode the server response. This can happen if the protocol changes in conjunction with strict json decoding.

Failure(Response)

Indicates an HTTP repsonse with a non-success status code